moved auth into a separate folder, added dev api key
