+++ alcf/config.py
@@ -4,4 +4,9 @@ import os
 from dotenv import load_dotenv
 load_dotenv()
 
+# Database configuration
 DATABASE_URL = os.getenv("DATABASE_URL", "postgresql+asyncpg://facilityapi_user@localhost/facilityapi_db")
+
+# Dev only
+# TODO: Remove this once Auth is integrated
+SECRET_DEV_KEY = os.getenv("SECRET_DEV_KEY", None)
